x86: Secure Launch SMP bringup support
On Intel, the APs are left in a well documented state after TXT performs the
late launch. Specifically they cannot have #INIT asserted on them so a standard
startup via INIT/SIPI/SIPI cannot be performed. Instead the early SL stub code
parked the APs in a pause/jmp loop waiting for an NMI. The modified SMP boot
code is called for the Secure Launch case. The jump address for the RM piggy
entry point is fixed up in the jump where the APs are waiting and an NMI IPI
is sent to the AP. The AP vectors to the Secure Launch entry point in the RM
piggy which mimics what the real mode code would do then jumps the the standard
RM piggy protected mode entry point.

For AMD things are quite different. The SKINIT instruction leaves the APs
halted with #INIT asserted. So the normal SMP boot code can just issue the
INIT/SIPI/SIPI sequence on the APs. The only change is that all interrupts are
globally disabled on AMD APs and must be re-enabled with the STGI instruction.
This is done in the entry to the RM piggy.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>