x86: Secure Launch kernel early boot stub
The Secure Launch (SL) stub provides the entry point for Intel TXT or
AMD SKINIT to vector to during the late launch. The symbol sl_stub_entry is
that entry point and is conveyed launching code using the MLE (Measured Launch
Environment) header represented by the symbol sl_mle_header. The offset of the
MLE is written to the boot params at buld time. The routine sl_stub contains
the very early late launch setup code responsible for setting up the basic
environment to allow the normal kernel startup_32 code to proceed. It is also
responsible for properly waking and handling the APs on Intel platforms. The
routine sl_main which runs after entering 64b mode is responsible for measuring
configuration information before it is used like the boot params, the kernel
command line, the TXT heap, etc.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>