x86: Secure Launch kernel late boot stub
The routine slaunch_setup is called out of the x86 specific setup_arch
routine during early kernel boot. After determining what platform is present,
various operations specific to that platform occur. This includes finalizing
setting for the platform late launch and verifying that memory protections
are in place.

For TXT, this code also reserves the original compressed kernel setup area
where the APs were left looping so that this memory cannot be used.

Signed-off-by: Ross Philipson <ross.philipson@oracle.com>