x86: Secure Launch adding event log securityfs
The late init functionality registers securityfs nodes to allow fetching
of the late launch TPM log and other platform specific operations. TXT also
stores a copy of the DMAR ACPI table to protect it from tampering during the
late launch. This copy of the table will be made available to the ACPI drivers
instead of the one found in the firmware.

Signed-off-by: Daniel P. Smith <dpsmith@apertussolutions.com>